requirement.txt
#python==3.10
#conda-forge::openslide
timm==0.9.8
torch
torchvision
h5py
pandas
PyYAML
opencv-python
matplotlib
scikit-learn
scipy
tqdm
openslide-python
git+https://github.com/oval-group/smooth-topk.git
tensorboardX
